+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Raising Error objects
 * Decreased $mui-container-sm width by 18px to account for vertical scrollbar
 * Bugfix to _outsideClick handler in react library
+* Using default <select> behavior on touch devices
 
 ## 0.1.19 - July 26, 2015